Scripting API: make MemoryBreakpoint value u64
This fixes a follow-up of 7a6b63b309cd1dac4f402641b9d80c646442638e "MMU: Don't truncate 64-bit values when calling Memcheck()".

+++ Source/Core/Core/API/Events.h
@@ -21,7 +21,7 @@ struct MemoryBreakpoint
 {
   bool write;
   u32 addr;
-  u32 value;
+  u64 value;
 };
 struct SetInterrupt
 {
